[feat] 인기메뉴 목록 조회 기능 추가

+++ src/main/java/com/hyun/CoffeOrderingSystem/dto/MenuInfoResp.java
@@ -7,9 +7,9 @@ import java.math.BigDecimal;
 public class MenuInfoResp {
     private Long menuId;
     private String name;
-    private BigDecimal price;
+    private Long price;
 
-    private MenuInfoResp(Long menuId, String name, BigDecimal price) {
+    private MenuInfoResp(Long menuId, String name, Long price) {
         this.menuId = menuId;
         this.name = name;
         this.price = price;
